Revert "SALTO-2768 align the workflow of PermissionSet to Profile  (#3483)"
This reverts commit 1a81d18f56166f19d394a1970ad3b1011f56af4b.
