+++ packages/salesforce-adapter/src/constants.ts
@@ -295,7 +295,7 @@ export const DEFAULT_MAX_INSTANCES_PER_TYPE = 5000
 export const MINIMUM_MAX_ITEMS_IN_RETRIEVE_REQUEST = 500
 export const MAXIMUM_MAX_ITEMS_IN_RETRIEVE_REQUEST = 10000
 export const MAX_QUERY_LENGTH = 2000
-export const DEFAULT_ENUM_FIELD_PERMISSIONS = true
+export const DEFAULT_ENUM_FIELD_PERMISSIONS = false
 export const DEFAULT_CUSTOM_OBJECTS_DEFAULT_RETRY_OPTIONS = {
   maxAttempts: 3,
   retryDelay: 1000,
