+++ packages/salesforce-adapter/src/types.ts
@@ -84,7 +84,7 @@ export type ChangeValidatorName = (
   | 'unknownField'
   | 'customFieldType'
   | 'standardFieldLabel'
-  | 'mapKeys'
+  | 'profileMapKeys'
   | 'multipleDefaults'
   | 'picklistPromote'
   | 'cpqValidator'
@@ -549,7 +549,7 @@ const changeValidatorConfigType = createMatchingObjectType<ChangeValidatorConfig
     unknownField: { refType: BuiltinTypes.BOOLEAN },
     customFieldType: { refType: BuiltinTypes.BOOLEAN },
     standardFieldLabel: { refType: BuiltinTypes.BOOLEAN },
-    mapKeys: { refType: BuiltinTypes.BOOLEAN },
+    profileMapKeys: { refType: BuiltinTypes.BOOLEAN },
     multipleDefaults: { refType: BuiltinTypes.BOOLEAN },
     picklistPromote: { refType: BuiltinTypes.BOOLEAN },
     cpqValidator: { refType: BuiltinTypes.BOOLEAN },
