+++ packages/salesforce-adapter/test/filters/field_permissions_enum.test.ts
@@ -13,45 +13,15 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-import {
-  InstanceElement,
-  isMapType,
-  getDeepInnerType,
-  ObjectType,
-  ElemID,
-  MapType,
-  toChange,
-  getChangeData,
-  Change,
-  ListType,
-} from '@salto-io/adapter-api'
+import { InstanceElement, isMapType, getDeepInnerType, ObjectType, ElemID, MapType, toChange, getChangeData, Change } from '@salto-io/adapter-api'
 import { FilterWith } from '../../src/filter'
 import * as constants from '../../src/constants'
 import fieldPermissionsEnumFilter, { enumFieldPermissions, profileFieldLevelSecurity } from '../../src/filters/field_permissions_enum'
 import { generateProfileType, defaultFilterContext } from '../utils'
-import { METADATA_TYPE, PERMISSION_SET_METADATA_TYPE, SALESFORCE } from '../../src/constants'
 
 describe('FieldPermissionsEnum filter', () => {
   let filter: FilterWith<'onFetch' | 'onDeploy' | 'preDeploy'>
   const profileObj = generateProfileType(true)
-  const permissionSetObject = new ObjectType({
-    elemID: new ElemID(SALESFORCE, PERMISSION_SET_METADATA_TYPE),
-    fields: {
-      fieldPermissions: { refType: new MapType(new MapType(profileFieldLevelSecurity)) },
-    },
-    annotations: {
-      [METADATA_TYPE]: PERMISSION_SET_METADATA_TYPE,
-    },
-  })
-  const permissionSetObjectBeforeCovnert = new ObjectType({
-    elemID: new ElemID(SALESFORCE, PERMISSION_SET_METADATA_TYPE),
-    fields: {
-      fieldPermissions: { refType: new ListType(profileFieldLevelSecurity) },
-    },
-    annotations: {
-      [METADATA_TYPE]: PERMISSION_SET_METADATA_TYPE,
-    },
-  })
   const fieldPermissionObjectValue = {
     ObjA: {
       FieldA: {
@@ -88,18 +58,6 @@ describe('FieldPermissionsEnum filter', () => {
       },
     },
   }
-  const fieldPermissionObjectValueAsList = [
-    {
-      field: 'ObjA.FieldA',
-      editable: true,
-      readable: true,
-    },
-    {
-      field: 'ObjA.FieldB',
-      editable: false,
-      readable: true,
-    },
-  ]
   const fieldPermissionEnumValue = {
     ObjA: {
       FieldA: 'ReadWrite',
@@ -119,43 +77,18 @@ describe('FieldPermissionsEnum filter', () => {
       fieldPermissions: fieldPermissionObjectValue,
     }
   )
-  const permissionSetInstance = new InstanceElement(
-    'permissionSetInst',
-    permissionSetObject,
-    {
-      fieldPermissions: fieldPermissionObjectValue,
-    }
-  )
-  const permissionSetInstanceBeforeConvert = new InstanceElement(
-    'permissionSetInstanceBeforeConvert',
-    permissionSetObjectBeforeCovnert, {
-      fieldPermissions: fieldPermissionObjectValueAsList,
-    }
-  )
 
   describe('onFetch', () => {
     let elements: (InstanceElement | ObjectType)[]
     let profileInstanceClone: InstanceElement
-    let permissionSetInstanceClone: InstanceElement
-    let permissionSetObjectClone: ObjectType
     let profileObjectClone: ObjectType
-    let permissionSetbjectBeforeConvertClone: ObjectType
-    let permissionSetInstanceBeforeConvertClone: InstanceElement
     describe('with enumFieldPermissions true', () => {
       beforeAll(async () => {
         profileInstanceClone = profileInstance.clone()
-        permissionSetInstanceClone = permissionSetInstance.clone()
-        permissionSetObjectClone = permissionSetObject.clone()
         profileObjectClone = profileObj.clone()
-        permissionSetbjectBeforeConvertClone = permissionSetObjectBeforeCovnert.clone()
-        permissionSetInstanceBeforeConvertClone = permissionSetInstanceBeforeConvert.clone()
         elements = [
           profileObjectClone,
-          permissionSetObjectClone,
           profileInstanceClone,
-          permissionSetInstanceClone,
-          permissionSetbjectBeforeConvertClone,
-          permissionSetInstanceBeforeConvertClone,
         ]
         filter = fieldPermissionsEnumFilter(
           { config: { ...defaultFilterContext, enumFieldPermissions: true } },
@@ -172,28 +105,10 @@ describe('FieldPermissionsEnum filter', () => {
           .toBeTruthy()
       })
 
-      it('Should convert PermissionSet Object fieldPermissions type to fieldPermissionEnum', async () => {
-        const fieldPermissionsFieldType = await (permissionSetObjectClone)
-          .fields.fieldPermissions.getType()
-        expect(isMapType(fieldPermissionsFieldType)).toBeTruthy()
-        const deepInnerFieldPermissionType = await getDeepInnerType(fieldPermissionsFieldType)
-        expect(deepInnerFieldPermissionType.elemID.isEqual(enumFieldPermissions.elemID))
-          .toBeTruthy()
-      })
-
-      it('Should convert Profile and PermissionSet instances\' fieldPermissions values to right enums', async () => {
-        [profileInstanceClone, permissionSetInstanceClone].forEach(instance => {
-          expect(instance.value).toEqual({
-            fieldPermissions: fieldPermissionEnumValue,
-          })
+      it('Should convert Profile instances\' fieldPermissions values to right enums', async () => {
+        expect(profileInstanceClone.value).toEqual({
+          fieldPermissions: fieldPermissionEnumValue,
         })
-      })
-      it('Should not covert Profile and PermissionSet type and instances if field did not convert to map', async () => {
-        expect(permissionSetInstanceBeforeConvertClone.value.fieldPermissions)
-          .toEqual(fieldPermissionObjectValueAsList)
-        expect(isMapType(
-          await permissionSetbjectBeforeConvertClone.fields.fieldPermissions.getType()
-        )).toBeFalsy()
       })
     })
 
@@ -203,40 +118,25 @@ describe('FieldPermissionsEnum filter', () => {
           { config: { ...defaultFilterContext } },
         ) as FilterWith<'onFetch' | 'onDeploy' | 'preDeploy'>
         profileInstanceClone = profileInstance.clone()
-        permissionSetInstanceClone = permissionSetInstance.clone()
-        permissionSetObjectClone = permissionSetObject.clone()
         profileObjectClone = profileObj.clone()
         elements = [
           profileObjectClone,
-          permissionSetObjectClone,
           profileInstanceClone,
-          permissionSetInstanceClone,
         ]
         await filter.onFetch(elements)
       })
 
-      it('Should not change the Profile and PermissionSet objects', async () => {
+      it('Should not change the Profile objects', async () => {
         expect(profileObj.isEqual(profileObjectClone)).toBeTruthy()
-        expect(permissionSetObject.isEqual(permissionSetObjectClone)).toBeTruthy()
       })
 
-      it('Should not change Profile and PermissionSet instances', () => {
+      it('Should not change Profile instances', () => {
         expect(profileInstance.isEqual(profileInstanceClone)).toBeTruthy()
-        expect(permissionSetInstance.isEqual(permissionSetInstanceClone)).toBeTruthy()
       })
     })
   })
 
   describe('deploy (pre + on)', () => {
-    const permissionSetObjectPostOnFetch = new ObjectType({
-      elemID: new ElemID(SALESFORCE, PERMISSION_SET_METADATA_TYPE),
-      fields: {
-        fieldPermissions: { refType: new MapType(new MapType(enumFieldPermissions)) },
-      },
-      annotations: {
-        [METADATA_TYPE]: PERMISSION_SET_METADATA_TYPE,
-      },
-    })
     const profileObjectPostOnFetch = new ObjectType({
       elemID: new ElemID(constants.SALESFORCE, constants.PROFILE_METADATA_TYPE),
       fields: {
@@ -253,13 +153,6 @@ describe('FieldPermissionsEnum filter', () => {
         fieldPermissions: fieldPermissionEnumValue,
       }
     )
-    const permissionSetInstancePostOnFetch = new InstanceElement(
-      'permissionSetInstPostOnFetch',
-      permissionSetObjectPostOnFetch,
-      {
-        fieldPermissions: fieldPermissionEnumValue,
-      }
-    )
     let changes: Change<InstanceElement>[]
 
     describe('with instances and types that had this filter\'s onFetch run on them', () => {
@@ -269,7 +162,6 @@ describe('FieldPermissionsEnum filter', () => {
         ) as FilterWith<'onFetch' | 'onDeploy' | 'preDeploy'>
         changes = [
           profileInstancePostOnFetch,
-          permissionSetInstancePostOnFetch,
         ].map(elem => toChange({ after: elem.clone() }))
       })
 
@@ -329,7 +221,6 @@ describe('FieldPermissionsEnum filter', () => {
         ) as FilterWith<'onFetch' | 'onDeploy' | 'preDeploy'>
         changes = [
           profileInstance,
-          permissionSetInstance,
         ].map(elem => toChange({ after: elem.clone() }))
       })
